Pass list of objects to turn()
Now creatures can't move on top of each other, and this is necessary
otherwise for obvious reasons.

So. I needed to loop through `objects' and have a mutable reference to
each item inside of it while at the same time having a mutable reference
to the rest of the vector.

Obviously you can't have two mut refs to the same thing; therefore, I
have to use the split_at_mut family of functions.

However! It needs to work for all array elements, which means for
elements in the middle, "the rest of the Vec" is not contiguous. and
therefore cannot be split off cleanly to get a separate mutable
reference to.

Which means I have to split once *before* the element in question, and
then split the right half again.

Which means I have to pass *two* references to turn(), which processes
"the rest of the vector"--one for the first part and one for the second.

Which is just ugly, ugh.

</rant>